+++ src/main/java/com/jovan/erp_v1/model/StockTransferItem.java
@@ -1,5 +1,7 @@
 package com.jovan.erp_v1.model;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -24,9 +26,8 @@ public class StockTransferItem {
     private Product product;
 
     @Column
-    private Double quantity;
+    private BigDecimal quantity;
 
     @ManyToOne
     private StockTransfer stockTransfer;
-
 }
