+++ src/main/java/com/jovan/erp_v1/request/StockTransferItemRequest.java
@@ -1,10 +1,12 @@
 package com.jovan.erp_v1.request;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record StockTransferItemRequest(
         Long id,
         @NotNull Long productId,
-        @NotNull @Positive Double quantity) {
+        @NotNull @Positive BigDecimal quantity) {
 }
